backend/requirements.txt
# Django & REST Framework
django==4.2.8
djangorestframework==3.14.0
dj-rest-auth==5.0.1
django-cors-headers==4.3.1
django-filter==24.1
django-crontab==0.7.1  # For scheduled tasks

# Database
psycopg2-binary==2.9.9

# Security & Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1

# Utilities
python-dotenv==1.0.1
Pillow==10.1.0  # For image processing
boto3==1.34.44  # For AWS S3 integration
django-storages==1.14.2  # For storage backends

# Export functionality 
openpyxl==3.1.2  # For Excel export
pandas==2.1.4  # For data manipulation
xlsxwriter==3.1.9  # For Excel writing

# Development tools
django-debug-toolbar==4.3.0
black==24.1.1  # Code formatting
isort==5.13.2  # Import sorting